did assignment

+++ c3prj1_deck/deck.c
@@ -113,8 +113,8 @@ deck_t * build_remaining_deck(deck_t ** hands, size_t n_hands){
   deck_t *tober=malloc(sizeof(*tober));
   tober->n_cards=0;
   tober->cards = NULL;
-  for(int i=0; i<n_hands; i++){
-    for(int j=0; j<hands[i]->n_cards; j++){
+  for(size_t i=0; i<n_hands; i++){
+    for(size_t j=0; j<hands[i]->n_cards; j++){
       add_card_to(tober, *(hands[i]->cards[j]));
     }
   }
